refactor: move the wall-clock helpers out of testing.rs
# Summary

`since_epoch` and `since_epoch_millis` now live in `utils`, so no
production code imports from a module named `testing`.

# Details

Nothing in `testing.rs` was test-only: both functions read the system
clock and `kv_api` calls `since_epoch_millis` on every expire-aware
read. The name suggested the opposite and hid a wall-clock dependency
in what looked like test scaffolding.

`utils` already collects the crate's small helpers, so the file is gone
rather than renamed. Both functions stay `pub(crate)`.
